refactor(rql): move AST→SQL renderer into reddb-io-rql with parser tests
The parser family's embedded property round-trip test
(render(parse(render(ast)))) must build as a single in-crate type
instance — routing render() through the reddb-server dev-dependency
pulled in a second AST copy and failed with a spurious type mismatch.
Move renderer.rs into reddb-io-rql (its only external callers are the
runtime's render_value_sql, preserved via the server shim) and widen
render_value_sql to pub for the cross-crate shim. Add serde_json +
proptest dev-deps for the relocated tests; mirror the source crate's
clippy allows so the lib stays lint-clean under the same posture.

Refs #1103
