+++ crates/reddb-rql/Cargo.toml
@@ -30,6 +30,9 @@ reddb-types = { package = "reddb-io-types", version = "1.0", path = "../reddb-ty
 reddb-server = { package = "reddb-io-server", version = "1.10", path = "../reddb-server" }
 sqllogictest = "0.29"
 # Parser property round-trip tests (issue #87) moved in with the parser family
-# (#1103); they reach the server-resident AST→SQL renderer via the
-# `reddb-server` dev-dependency above.
+# (#1103); the AST→SQL renderer moved with them, so the round-trip stays a
+# single in-crate type instance (no `reddb-server` round-trip).
 proptest = "1"
+# Embedded parser tests assert on parsed SECRET_REF JSON via the upstream
+# `serde_json` crate (the server carries the same dependency).
+serde_json = "1"
